requirement.txt
torch
torchvision
torchaudio
mne
numpy
pyyaml
tqdm
scipy
scikit-learn